增加了新增post和evnet后的自动刷新，更改了home页的布局，修改完成了引导页，完善了宿舍缴费页面，增加了个性签名修改、点击头像查看个人信息，post的长按删除

+++ app/src/main/java/com/stellaris/practice/MapActivity.java
@@ -301,7 +301,7 @@ public class MapActivity extends BaseMapActivity implements SensorEventListener,
                 commit_bt.setBackgroundDrawable(getResources().getDrawable(R.mipmap.restaurant_btbg_gray));
                 mBaiduMap.setMyLocationEnabled(true);
             }
-            mDistance_tv.setText("距离目的地：" + mDistance + "米");
+            mDistance_tv.setText("距离目的地：" + (int)mDistance + "米");
             //缩放地图
             setMapZoomScale(LocationPoint);
         }
